// src/types/anime.ts
export type Anime = {
    id: number;
    uuid?: string;
    title: string;
    description?: string;
    image_url: string;
    large_image_url?: string;
    small_image_url?: string;
    small_image_url_webp?: string;
    poster_url?: string;
    status?: string;
    release_date?: string;
    isPined?: boolean;
  };